test_nat.c: silenced warning.
darcs-hash:20061213143756-65a35-db23c48a319259e37be3f48fe54b07e3f1242911.gz

+++ libsofia-sip-ua/nua/test_nat.c
@@ -384,7 +384,7 @@ int test_nat_public(struct nat *nat, void const *address, int addrlen)
     return 0;
   }
 
-  if (addrlen > sizeof nat->out_address)
+  if ((size_t)addrlen > sizeof nat->out_address)
     return su_seterrno(EINVAL);
 
   for (li = nat->localinfo; li; li = li->li_next) {
